operations/color/brightness-contrast.c operations/color/whitebalance.c
2007-06-12  Mukund Sivaraman  <muks@mukund.org>

        * operations/color/brightness-contrast.c
        * operations/color/whitebalance.c
        * operations/color/invert.c
        * operations/color/levels.c
        * operations/color/threshold.c: Use glong instead of gint.


svn path=/trunk/; revision=1561

+++ operations/color/threshold.c
@@ -47,7 +47,7 @@ process (GeglOperation *op,
          void          *out_buf,
          glong          n_pixels)
 {
-  gint i;
+  glong   i;
 
   gfloat *in = in_buf;
   gfloat *out = out_buf;
